refactor(invalid-interactive): rename to no-invalid-interactive
