+++ docs/rule/no-invalid-interactive.md
@@ -1,4 +1,4 @@
-## invalid-interactive
+## no-invalid-interactive
 
 Adding interactivity to an element that is not naturally interactive content leads to a very poor experience for
 users of assistive technology (i.e. screen readers). In order to ensure that screen readers can provide useful information
