+++ lib/config/recommended.js
@@ -10,7 +10,7 @@ module.exports = {
     'self-closing-void-elements': true,
     'triple-curlies': true,
     'link-rel-noopener': true,
-    'invalid-interactive': true,
+    'no-invalid-interactive': true,
     'img-alt-attributes': true,
     'style-concatenation': true,
     'deprecated-inline-view-helper': true,
