+++ lib/rules/index.js
@@ -20,7 +20,7 @@ module.exports = {
   'deprecated-each-syntax': require('./deprecations/lint-deprecated-each-syntax'),
   'deprecated-inline-view-helper': require('./deprecations/lint-deprecated-inline-view-helper'),
   'deprecated-render-helper': require('./deprecations/lint-deprecated-render-helper'),
-  'invalid-interactive': require('./lint-invalid-interactive'),
+  'no-invalid-interactive': require('./lint-no-invalid-interactive'),
   'style-concatenation': require('./lint-style-concatenation'),
   'simple-unless': require('./lint-simple-unless'),
   'unused-block-params': require('./lint-unused-block-params'),
