+++ lib/rules/lint-no-invalid-interactive.js
